Move lib/utils/host to session/host
